% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-fit-new.R, R/Model-fit.R
\name{Fitted.model}
\alias{Fitted.model}
\title{Title}
\usage{
Fitted.model(
  Dataset,
  intercept = T,
  scale.lasso = T,
  f = linear,
  g = linear,
  sd.epsilon = sd.epsilon,
  C = 0.75,
  sd.v = sd.v,
  cvnfolds = 10
)

Fitted.model(
  Dataset,
  intercept = T,
  scale.lasso = T,
  f = linear,
  g = linear,
  sd.epsilon = sd.epsilon,
  C = 0.75,
  sd.v = sd.v,
  cvnfolds = 10
)
}
\arguments{
\item{Dataset}{}

\item{intercept}{}

\item{scale.lasso}{}

\item{f}{}

\item{g}{}

\item{sd.epsilon}{}

\item{C}{}

\item{sd.v}{}

\item{cvnfolds}{}
}
\value{



}
\description{
Title

Title
}
\examples{


}
