% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GECV.EST.MAIN}
\alias{GECV.EST.MAIN}
\title{Function for estimating the Genetic Covariance}
\usage{
GECV.EST.MAIN(
  Dataset,
  type = "split",
  intercept = T,
  scale.lasso = T,
  f = linear,
  g = linear,
  sd.epsilon = NULL,
  sd.v = NULL
)
}
\arguments{
\item{Dataset}{the \code{Dataset} argument should be a n * (p + 2) input matrix and each row is an observation vector. First two columns are two interested outcomes where missing values are allowed.
The last p columns are collected genotypes matrix for n individuals}

\item{type}{1-character string giving the type of "split" for sample splitting and cross-fitting procedure is implemented ; "whole" for using whole samples for both fitting the data and construct the confidence interval.}

\item{intercept}{Should intercept be fitted (default=TRUE) or set to zero (FALSE)?}

\item{scale.lasso}{Should the scale lasso or glmnet be used?}

\item{f}{Specified function form f(x_i beta) for outcome 1. 'linear' for linear model and 'logit' for logistic model of binary outcomes}

\item{g}{Specified function form g(x_i gamma) for outcome 2 . 'linear' for linear model and 'logit' for logistic model of binary outcomes}

\item{sd.epsilon, sd.v}{prior knownlege on the standard deviation of error terms epsilon and v}
}
\value{
A vector of two elements: estimated genetic covariance and standard errors
}
\description{
Function for estimating the Genetic Covariance
}
\examples{


}
