% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{GECV.EST}
\alias{GECV.EST}
\title{Title}
\usage{
GECV.EST(
  Dataset,
  beta.hat,
  gamma.hat,
  f,
  g,
  case.control = F,
  p1 = NULL,
  p0 = NULL
)
}
\arguments{
\item{Dataset}{}

\item{beta.hat}{}

\item{gamma.hat}{}

\item{f}{}

\item{g}{}

\item{case.control}{}

\item{p1}{}

\item{p0}{}
}
\value{

}
\description{
Title
}
\examples{

}
