% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ecv.R
\name{geom_ecv}
\alias{geom_ecv}
\title{geom_ecv}
\usage{
geom_ecv(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge ecvs}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer adding ecvs
}
\examples{
library("ggplot2")
ggplot(mtcars) +
geom_ecv(aes(mpg, wt), ecv = "bv", size = 5)

set.seed(1)
df <- data.frame(x = rnorm(10),
                 y = rnorm(10),
                 image = sample(c("bv",
                                  "bp",
                                  "mcv1",
                                  "pl",
                                  "sl"),
                                 size = 10, replace = TRUE))
 ggplot(df) +
geom_ecv(aes(x, y, ecv = image), size = 5)

}
