% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\docType{class}
\name{Regressor-class}
\alias{Regressor-class}
\alias{Regressor}
\title{Functions for generating data and coeficient vectors}
\value{
a matrix of data
}
\description{
A class for generating regressor matrix
}
\details{
... Complete documentation ...
}
\section{Fields}{

\describe{
\item{\code{n}}{numeric.}

\item{\code{p}}{numeric.}

\item{\code{Cov}}{matrix.}
}}


\examples{

}
